+++ addOns/authhelper/CHANGELOG.md
@@ -12,6 +12,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Include the Zest statement index in the authentication diagnostics' steps.
 - Include Origin header when doing authentication verification.
 
+### Fixed
+- Handle exception while extracting session tokens.
+
 ## [0.40.0] - 2026-06-12
 ### Added
 - Support for handling the Microsoft login "Permissions requested" user consent screen.